+++ lib/Bitbucket/API/Repositories/Pipelines.php
@@ -47,7 +47,9 @@ class Pipelines extends Api
     public function create($account, $repo, $params = array())
     {
         // allow developer to directly specify params as json if (s)he wants.
-        if ('array' !== gettype($params)) {
+        if (!is_array($params)) {
+            trigger_deprecation('private-packagist/bitbucket-api', '2.2', 'Calling Pipelines::create() with a string as params argument is deprecated. Pass an array instead.');
+
             if (empty($params)) {
                 throw new \InvalidArgumentException('Invalid JSON provided.');
             }